pyproject.toml
[tool.poetry]
name = "htmlextradata"
version = "0.1.2"
description = "MkDocs Plugin"
authors = ["Alexandr Starovoytov <stalex.info@yandex.ru>"]
repository = "https://github.com/stalexsm/mkdocs-htmlextradata-plugin"
readme = "README.md"
license = "MIT"
keywords = [
    "mkdocs", "mkdocs-plugin",
]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]

[project.urls]
Homepage = "https://github.com/stalexsm/mkdocs-htmlextradata-plugin"

[tool.poetry.dependencies]
python = "^3.10"
mkdocs = "^1.4.0"
pyyaml = "^6.0"

[tool.poetry.plugins."mkdocs.plugins"]
htmlextradata = "htmlextradata:HtmlExtraDataPlugin"

[tool.poetry.dev-dependencies]
pytest = "7.2.2"
black = "23.1.0"

[tool.black]
line-length = 88
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | \.env
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip=[".gitignore",".dockerignore", ".env", ".venv"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
